Split commit-section-list.tsx into per-component files
- Move CommitFileDetailContent / CommitItem / CommitList / CommitLoadMoreButton
  into their own files under components/commit-section/
- Keep CommitFileRow and CommitFileRows together (tightly coupled: Rows only
  maps files to Row) and keep buildCommitFilesSection colocated with
  CommitExpandedSection (its sole caller)
- Extract shared isKeyboardActivationKey into commit-section/commit-keyboard.ts
- Update CommitSection.tsx imports; delete the old aggregate file
