+++ apps/web/src/pages/SessionDetail/components/commit-section/commit-keyboard.ts
@@ -1,0 +1,4 @@
+import type { KeyboardEvent as ReactKeyboardEvent } from "react";
+
+export const isKeyboardActivationKey = (event: ReactKeyboardEvent<HTMLElement>) =>
+  event.key === "Enter" || event.key === " ";
